Fix Filter behavior
At the beginning, I was under the belief that in the event of OR
clauses SQLite would create a single cursor and then repeatedly call
the xFilter method on it, to "increase its scope". This is incorrect, as
SQLite will actually perform the merging of results from OR clauses on
its own by creating multiple cursors and then merging their results.
This commit removes complexity by leveraging this behavior.